Implement Data item to reuse component for Saved boards
