+++ src/app/[locale]/dashboard/savedData/DataItem.tsx
@@ -1,0 +1,40 @@
+import DeleteOutline from '@mui/icons-material/DeleteOutline';
+import EditOutlined from '@mui/icons-material/EditOutlined';
+import IconButton from '@mui/material/IconButton';
+import ListItem from '@mui/material/ListItem';
+import ListItemText from '@mui/material/ListItemText';
+import { DataRow } from './types';
+import { useFormatter } from 'next-intl';
+
+import Box from '@mui/material/Box';
+
+type Props = {
+  data: DataRow;
+  onDelete: (dataRow: DataRow) => void;
+};
+
+export default function DataItem({ data, onDelete }: Props) {
+  const format = useFormatter();
+
+  return (
+    <ListItem
+      divider
+      secondaryAction={
+        <Box>
+          <IconButton size="small">
+            <EditOutlined fontSize="small" />
+          </IconButton>
+          <IconButton onClick={() => onDelete(data)} size="small">
+            <DeleteOutline fontSize="small" />
+          </IconButton>
+        </Box>
+      }
+    >
+      <ListItemText
+        primary={data.prompt}
+        secondary={format.relativeTime(new Date(data.date))}
+        sx={{ pr: 4 }}
+      />
+    </ListItem>
+  );
+}
