+++ src/app/[locale]/dashboard/savedData/SavedData.tsx
@@ -1,10 +1,10 @@
 import { NextIntlClientProvider } from 'next-intl';
 import TabSelector from './Selector';
 
-export default function SavedData(props: { history: React.ReactNode }) {
+export default function SavedData({ history }: { history: React.ReactNode }) {
   return (
     <NextIntlClientProvider>
-      <TabSelector history={props.history}></TabSelector>
+      <TabSelector history={history} />
     </NextIntlClientProvider>
   );
 }
