+++ src/app/[locale]/dashboard/savedData/Selector.tsx
@@ -4,7 +4,7 @@ import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-
+import List from '@mui/material/List';
 interface TabPanelProps {
   children?: React.ReactNode;
   index: number;
