+++ src/app/[locale]/dashboard/savedData/types.ts
@@ -1,0 +1,5 @@
+export type DataRow = {
+  id: number | string;
+  prompt: string;
+  date: Date | string;
+};
